+++ versions/1.21.10/src/client/java/schrumbo/schrumbohud/clickgui/categories/GeneralCategory.java
@@ -1,23 +1,12 @@
 package schrumbo.schrumbohud.clickgui.categories;
 
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.font.TextRenderer;
-import net.minecraft.client.gui.DrawContext;
-import net.minecraft.text.Text;
 import schrumbo.schrumbohud.SchrumboHUDClient;
-import schrumbo.schrumbohud.Utils.ChatUtils;
-import schrumbo.schrumbohud.Utils.RenderUtils;
-import schrumbo.schrumbohud.clickgui.ClickGuiScreen;
-import schrumbo.schrumbohud.clickgui.widgets.ButtonWidget;
 import schrumbo.schrumbohud.clickgui.widgets.ColorPickerWidget;
-import schrumbo.schrumbohud.clickgui.widgets.SliderWidget;
 import schrumbo.schrumbohud.clickgui.widgets.ToggleWidget;
 import schrumbo.schrumbohud.config.HudConfig;
 
 public class GeneralCategory extends Category {
 
-    private final MinecraftClient client = net.minecraft.client.MinecraftClient.getInstance();
-    private final TextRenderer textRenderer = client.textRenderer;
     private final HudConfig config = SchrumboHUDClient.config;
 
     public GeneralCategory() {
@@ -57,8 +46,7 @@ public class GeneralCategory extends Category {
                 .color(() -> config.guicolors.accent, config::setAccentColor)
                 .build();
         widgets.add(accentColorPicker);
-
-
+        updateWidgetPositions(startX, startY);
     }
 
 }