+++ versions/1.21.10/src/client/java/schrumbo/schrumbohud/clickgui/categories/PositionCategory.java
@@ -1,26 +1,17 @@
 package schrumbo.schrumbohud.clickgui.categories;
 
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.font.TextRenderer;
-import net.minecraft.client.gui.DrawContext;
-import net.minecraft.text.Text;
-import schrumbo.schrumbohud.SchrumboHUD;
 import schrumbo.schrumbohud.SchrumboHUDClient;
-import schrumbo.schrumbohud.Utils.RenderUtils;
 import schrumbo.schrumbohud.clickgui.widgets.ButtonWidget;
-import schrumbo.schrumbohud.clickgui.widgets.SliderWidget;
 import schrumbo.schrumbohud.config.ConfigManager;
 import schrumbo.schrumbohud.config.HudConfig;
 import schrumbo.schrumbohud.hud.HudEditorScreen;
 
 public class PositionCategory extends Category {
 
-    private final MinecraftClient client = net.minecraft.client.MinecraftClient.getInstance();
-    private final TextRenderer textRenderer = client.textRenderer;
-    private final HudConfig config = SchrumboHUDClient.config;
 
     public PositionCategory() {
-        super("Size and Position");
+        super("Position");
     }
 
     @Override
@@ -53,7 +44,7 @@ public class PositionCategory extends Category {
                 })
                 .build();
         widgets.add(changePos);
-
+        updateWidgetPositions(startX, startY);
     }
 
 }