+++ versions/1.21.10/src/client/java/schrumbo/schrumbohud/clickgui/categories/PresetsCategory.java
@@ -1,24 +1,12 @@
 package schrumbo.schrumbohud.clickgui.categories;
-
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.font.TextRenderer;
-import net.minecraft.client.gui.DrawContext;
-import net.minecraft.text.Text;
 import schrumbo.schrumbohud.SchrumboHUDClient;
-import schrumbo.schrumbohud.Utils.RenderUtils;
 import schrumbo.schrumbohud.clickgui.widgets.ButtonWidget;
-import schrumbo.schrumbohud.clickgui.widgets.ColorPickerWidget;
-import schrumbo.schrumbohud.clickgui.widgets.SliderWidget;
-import schrumbo.schrumbohud.clickgui.widgets.ToggleWidget;
 import schrumbo.schrumbohud.config.ConfigManager;
-import schrumbo.schrumbohud.config.HudConfig;
-import schrumbo.schrumbohud.hud.HudEditorScreen;
+
+
 
 public class PresetsCategory extends Category {
 
-    private final MinecraftClient client = net.minecraft.client.MinecraftClient.getInstance();
-    private final TextRenderer textRenderer = client.textRenderer;
-    private final HudConfig config = SchrumboHUDClient.config;
 
     public PresetsCategory() {
         super("Presets");
@@ -78,7 +66,7 @@ public class PresetsCategory extends Category {
                 })
                 .build();
         widgets.add(dracula);
-
+        updateWidgetPositions(startX, startY);
     }
 
 }