+++ versions/1.21.10/src/client/java/schrumbo/schrumbohud/clickgui/categories/TextCategory.java
@@ -1,20 +1,12 @@
 package schrumbo.schrumbohud.clickgui.categories;
 
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.font.TextRenderer;
-import net.minecraft.client.gui.DrawContext;
-import net.minecraft.text.Text;
 import schrumbo.schrumbohud.SchrumboHUDClient;
-import schrumbo.schrumbohud.Utils.RenderUtils;
 import schrumbo.schrumbohud.clickgui.widgets.ColorPickerWidget;
-import schrumbo.schrumbohud.clickgui.widgets.SliderWidget;
 import schrumbo.schrumbohud.clickgui.widgets.ToggleWidget;
 import schrumbo.schrumbohud.config.HudConfig;
 
 public class TextCategory extends Category {
 
-    private final MinecraftClient client = net.minecraft.client.MinecraftClient.getInstance();
-    private final TextRenderer textRenderer = client.textRenderer;
     private final HudConfig config = SchrumboHUDClient.config;
 
     public TextCategory() {
@@ -43,6 +35,7 @@ public class TextCategory extends Category {
                 .color(() -> config.colors.text, config::setTextColor)
                 .build();
         widgets.add(textColorPicker);
+        updateWidgetPositions(startX, startY);
     }
 
 }