+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # SchrumboHUD v1.0.5
-- **fixed no chat message being sent when togglling the hud on some versions**
-- **added support for 1.21.6 and 1.21.7**
+- **changed layout of the clickgui**
 
 # Known issues
 - **hud only being resizable on 1.21.5 -> disabled on the newer versions because of scaling issues with items**
-
+- **Text not being 100% centered on widgets**